refactor: Remove unecessary component's layer
A component that has no responsibility is not necessary
but the new App component will be chunked.
